Improve check_ratios.sh and scan_logs.sh
This commit improves both check_ratios.sh and scan_logs.sh.
It adds support in check_ratios.sh for the TMP_DIR environment
variable and --expr and --yes flags, and it also removes code
the reverses the order of lines in the log file.  Starting with
this commit, scan_logs.sh will produce log files in the oldest
to the newest order.

Changes were tested locally when analyzing the DoS issue.