Improve check_ratios.sh and scan_logs.sh (#30)
This commit improves both check_ratios.sh and scan_logs.sh.
It adds support in check_ratios.sh for the TMP_DIR environment
variable and --expr and --yes flags, and it also removes code
the reverses the order of lines in the log file.  Starting with
this commit, scan_logs.sh will produce log files in the oldest
to the newest order.

This commit also adds run_check_ratios.sh which is a wrapper
for check_ratios.sh.  This wrapper script simplifies the process
of running the check_ratios.sh with different configurations.

Changes were tested locally when analyzing the DoS issue.