Sync server's cloned repo with GitHub
This commit consolidates a large set of changes to bring
the server's cloned version of the repository in sync with
the GitHub version.  While it's generally best practice to
avoid large commits, this approach was discussed and agreed
upon internally, given that the code has been running
reliably in production.